+++ POSPRA.Domain/Entities/FileRecord.cs
@@ -3,7 +3,7 @@
     public class FileRecord
     {
         public int ID { get; set; }
-        public int POSID { get; set; }
+        public long POSID { get; set; }
         public string? InvoiceData { get; set; }
         public string? InvoiceNumber { get; set; }
         public int IsSynced { get; set; }
